refactor: move NodeContainers to another file
